And De-cdbs it too
Since switching to compat level 5 makes it explode into flames.

Put all the headers into the libsofia-sip-ua-dev package.  There is actually
only one header for the glib package, and that package depends on the -ua
package in any case, so just keep it simple and don't stuff around.

The tortuous list of headers for the -ua was wrong anyway, so just simplify
it to always install them all.

The sofia-sip-doc.install file not only pointed at the wrong (non-existing)
files, it also had completely the wrong name.  Make that sofia-sip-doc.docs
and include the -glib docs as well.

Drop the .la files.  telepathy-sofiasip is the only rdep, so let's just fix it
and be done with this today.

Drop the ${shlibs:Depends} for the -dev packages.  Symlinks don't usually have
symbols that need resolving.