Fix undefined behaviour accessing msg_mclass_s members
The original loop iterates over discrete msg_href_t members using the
mc_request member as a base.  But this isn't valid code, since the
struct may be padded between each member, and this indeed fails horribly
when compiled with optimisation using gcc 4.8.2-8.

Thanks to Radist Morse for pinning down the precise location of the buggy
code, and the patch posted here:
https://bugzilla.redhat.com/show_bug.cgi?id=981056